+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### Bugs Fixed
 
-* [#216](https://github.com/clojure-emacs/orchard/issues/216): Fix inspector navigation (`down` and `next-sibling`) with pagination.
+* [#216](https://github.com/clojure-emacs/orchard/issues/216): Fix some pagination issues with inspector navigation.
+  * `down` fn goes beyond current page. `next-sibling` and `previous-sibling` fns work in all pages.
 
 ## 0.23.0 (2024-03-03)